Don't 'declare global' in interface files
