+++ grafast/grafast/src/interfaces.ts
@@ -1,15 +1,10 @@
-import type LRU from "@graphile/lru";
 import type EventEmitter from "eventemitter3";
-import type { PluginHook } from "graphile-config";
 import type {
   ASTNode,
-  DocumentNode,
-  ExecutionArgs as GraphQLExecutionArgs,
   FragmentDefinitionNode,
   GraphQLArgs,
   GraphQLArgument,
   GraphQLArgumentConfig,
-  GraphQLError,
   GraphQLField,
   GraphQLFieldConfig,
   GraphQLInputField,
@@ -22,7 +17,6 @@ import type {
   GraphQLScalarType,
   GraphQLSchema,
   GraphQLType,
-  OperationDefinitionNode,
   ValueNode,
   VariableNode,
 } from "graphql";
@@ -42,8 +36,6 @@ import type {
   ConstantStep,
 } from "./steps/index.js";
 import type { GrafastInputObjectType, GrafastObjectType } from "./utils.js";
-
-type PromiseOrValue<T> = T | Promise<T>;
 
 export interface GrafastTimeouts {
   /**
@@ -114,202 +106,6 @@ export interface LinkedList<T> {
 }
 
 export const $$hooked = Symbol("hookArgsApplied");
-
-declare global {
-  namespace Grafast {
-    type ExecutionArgs = Pick<
-      GraphQLExecutionArgs,
-      "schema" | "document" | "rootValue" | "variableValues" | "operationName"
-    > & { [$$hooked]?: boolean; contextValue: Grafast.Context };
-
-    /**
-     * Details about the incoming GraphQL request - e.g. if it was sent over an
-     * HTTP request, the request itself so headers can be interrogated.
-     *
-     * It's anticipated this will be expanded via declaration merging, e.g. if
-     * your server is Koa v2 then `RequestContext.koav2` might be added.
-     */
-    interface RequestContext {}
-
-    /**
-     * The GraphQL context our schemas expect, generally generated from details in Grafast.RequestContext
-     */
-    interface Context {}
-
-    interface FieldExtensions {
-      plan?: FieldPlanResolver<any, any, any>;
-      subscribePlan?: FieldPlanResolver<any, any, any>;
-    }
-
-    interface ArgumentExtensions {
-      // fooPlan?: ArgumentPlanResolver<any, any, any, any, any>;
-      inputPlan?: ArgumentInputPlanResolver;
-      applyPlan?: ArgumentApplyPlanResolver;
-      autoApplyAfterParentPlan?: boolean;
-      autoApplyAfterParentSubscribePlan?: boolean;
-    }
-
-    interface InputObjectTypeExtensions {
-      inputPlan?: InputObjectTypeInputPlanResolver;
-    }
-
-    interface InputFieldExtensions {
-      // fooPlan?: InputObjectFieldPlanResolver<any, any, any, any>;
-      inputPlan?: InputObjectFieldInputPlanResolver;
-      applyPlan?: InputObjectFieldApplyPlanResolver;
-      autoApplyAfterParentInputPlan?: boolean;
-      autoApplyAfterParentApplyPlan?: boolean;
-    }
-
-    interface ObjectTypeExtensions {
-      assertStep?:
-        | ((step: ExecutableStep) => asserts step is ExecutableStep)
-        | { new (...args: any[]): ExecutableStep }
-        | null;
-    }
-
-    interface EnumTypeExtensions {}
-
-    interface EnumValueExtensions {
-      /**
-       * EXPERIMENTAL!
-       *
-       * @internal
-       */
-      applyPlan?: EnumValueApplyPlanResolver<any>;
-    }
-
-    interface ScalarTypeExtensions {
-      plan?: ScalarPlanResolver;
-      inputPlan?: ScalarInputPlanResolver;
-      /**
-       * Set true if `serialize(serialize(foo)) === serialize(foo)` for all foo
-       */
-      idempotent?: boolean;
-    }
-
-    interface SchemaExtensions {
-      /**
-       * Maximum number of queries to store in this schema's query cache.
-       */
-      queryCacheMaxLength?: number;
-
-      /**
-       * The underlying query cache
-       */
-      [$$queryCache]?: LRU<string, DocumentNode | ReadonlyArray<GraphQLError>>;
-
-      /**
-       * Maximum number of operations to store an operation plan lookup cache for
-       */
-      operationsCacheMaxLength?: number;
-
-      /**
-       * Maximum number of operation plans to store in a single operation's cache
-       */
-      operationOperationPlansCacheMaxLength?: number;
-
-      /**
-       * The starting point for finding/storing the relevant OperationPlan for a request.
-       */
-      [$$cacheByOperation]?: LRU<
-        OperationDefinitionNode,
-        CacheByOperationEntry
-      >;
-    }
-  }
-  namespace GraphileConfig {
-    interface GrafastOptions {
-      /**
-       * An object to merge into the GraphQL context. Alternatively, pass an
-       * (optionally asynchronous) function that returns an object to merge into
-       * the GraphQL context.
-       */
-      context?:
-        | Partial<Grafast.Context>
-        | ((
-            ctx: Partial<Grafast.RequestContext>,
-            args: Grafast.ExecutionArgs,
-          ) => PromiseOrValue<Partial<Grafast.Context>>);
-
-      /**
-       * A list of 'explain' types that should be included in `extensions.explain`.
-       *
-       * - `plan` will cause the plan JSON to be included
-       * - other values are dependent on the plugins in play
-       *
-       * If set to `true` then all possible explain types will be exposed.
-       */
-      explain?: boolean | string[];
-
-      timeouts?: GrafastTimeouts;
-    }
-    interface Preset {
-      /**
-       * Options that control how `grafast` should execute your GraphQL
-       * operations.
-       */
-      grafast?: GraphileConfig.GrafastOptions;
-    }
-    interface GrafastHooks {
-      args: PluginHook<
-        (event: {
-          args: Grafast.ExecutionArgs;
-          ctx: Grafast.RequestContext;
-          resolvedPreset: GraphileConfig.ResolvedPreset;
-        }) => PromiseOrValue<void>
-      >;
-    }
-    interface Plugin {
-      grafast?: {
-        hooks?: GrafastHooks;
-      };
-    }
-  }
-}
-
-/*
- * We register certain things (plans, etc) into the GraphQL "extensions"
- * property on the various GraphQL configs (type, field, argument, etc); this
- * uses declaration merging so that these can be accessed with types.
- */
-declare module "graphql" {
-  interface GraphQLFieldExtensions<_TSource, _TContext, _TArgs = any> {
-    grafast?: Grafast.FieldExtensions;
-  }
-
-  interface GraphQLArgumentExtensions {
-    grafast?: Grafast.ArgumentExtensions;
-  }
-
-  interface GraphQLInputObjectTypeExtensions {
-    grafast?: Grafast.InputObjectTypeExtensions;
-  }
-
-  interface GraphQLInputFieldExtensions {
-    grafast?: Grafast.InputFieldExtensions;
-  }
-
-  interface GraphQLObjectTypeExtensions<_TSource = any, _TContext = any> {
-    grafast?: Grafast.ObjectTypeExtensions;
-  }
-
-  interface GraphQLEnumTypeExtensions {
-    grafast?: Grafast.EnumTypeExtensions;
-  }
-
-  interface GraphQLEnumValueExtensions {
-    grafast?: Grafast.EnumValueExtensions;
-  }
-
-  interface GraphQLScalarTypeExtensions {
-    grafast?: Grafast.ScalarTypeExtensions;
-  }
-
-  interface GraphQLSchemaExtensions {
-    grafast?: Grafast.SchemaExtensions;
-  }
-}
 
 export const $$grafastContext = Symbol("context");
 export const $$planResults = Symbol("planResults");
